# content-builder-web-app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urgent Care API",
    "description": "Urgent Care API — content CMS, learning stats, auth, media. OpenAPI: `/openapi.json` · Swagger UI: `/docs` · ReDoc: `/redoc`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/billing/tariffs": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "List Tariffs",
        "operationId": "billing_list_tariffs",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TariffOut"
                  },
                  "type": "array",
                  "title": "Response Billing List Tariffs"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Tariff",
        "operationId": "billing_create_tariff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TariffCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/tariffs/all": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "List Tariffs Admin",
        "operationId": "billing_list_tariffs_admin",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TariffOut"
                  },
                  "type": "array",
                  "title": "Response Billing List Tariffs Admin"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/tariffs/{tariff_id}": {
      "patch": {
        "tags": [
          "billing"
        ],
        "summary": "Update Tariff",
        "operationId": "billing_update_tariff",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tariff Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TariffUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "billing"
        ],
        "summary": "Delete Tariff",
        "operationId": "billing_delete_tariff",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tariff Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/me": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Billing Me",
        "operationId": "billing_billing_me",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingMeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/subscribe": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Subscribe",
        "operationId": "billing_subscribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscribeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/subscription/cancel": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Cancel Subscription",
        "operationId": "billing_cancel_subscription",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingMeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/payments": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "List Payments",
        "operationId": "billing_list_payments",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PaymentOut"
                  },
                  "type": "array",
                  "title": "Response Billing List Payments"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/payments/{payment_id}/sync": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Sync Payment",
        "operationId": "billing_sync_payment",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/folders": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "List Folders",
        "operationId": "folders_list_folders",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parentId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parentid"
            }
          },
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "All"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderOut"
                  },
                  "title": "Response Folders List Folders"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "folders"
        ],
        "summary": "Create Folder",
        "operationId": "folders_create_folder",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/folders/{item_id}": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "Get Folder",
        "operationId": "folders_get_folder",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "folders"
        ],
        "summary": "Update Folder",
        "operationId": "folders_update_folder",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "folders"
        ],
        "summary": "Delete Folder",
        "operationId": "folders_delete_folder",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingMeOut": {
        "properties": {
          "tariffId": {
            "type": "string",
            "format": "uuid",
            "title": "Tariffid"
          },
          "tariffCode": {
            "type": "string",
            "title": "Tariffcode"
          },
          "tariffTitle": {
            "type": "string",
            "title": "Tarifftitle"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "currentPeriodStart": {
            "type": "string",
            "format": "date-time",
            "title": "Currentperiodstart"
          },
          "currentPeriodEnd": {
            "type": "string",
            "format": "date-time",
            "title": "Currentperiodend"
          },
          "cancelAtPeriodEnd": {
            "type": "boolean",
            "title": "Cancelatperiodend"
          },
          "priceRub": {
            "type": "integer",
            "title": "Pricerub"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "enforcement": {
            "type": "boolean",
            "title": "Enforcement"
          },
          "scheduledTariffId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledtariffid"
          },
          "scheduledTariffCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledtariffcode"
          },
          "scheduledTariffTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledtarifftitle"
          },
          "scheduledEffectiveAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledeffectiveat"
          },
          "scheduledChangeStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledchangestatus"
          }
        },
        "type": "object",
        "required": [
          "tariffId",
          "tariffCode",
          "tariffTitle",
          "status",
          "currentPeriodStart",
          "currentPeriodEnd",
          "cancelAtPeriodEnd",
          "priceRub",
          "rank",
          "enforcement"
        ],
        "title": "BillingMeOut"
      },
...
      "FolderOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid"
          },
          "requiredTariffId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiredtariffid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "order",
          "parentId",
          "requiredTariffId"
        ],
        "title": "FolderOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tariffId": {
            "type": "string",
            "format": "uuid",
            "title": "Tariffid"
          },
          "amountRub": {
            "type": "number",
            "title": "Amountrub"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "yookassaPaymentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yookassapaymentid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tariffId",
          "amountRub",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "PaymentOut"
      },
...
      "SubscribeOut": {
        "properties": {
          "confirmationUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmationurl"
          },
          "paymentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paymentid"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "default": false
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "scheduled": {
            "type": "boolean",
            "title": "Scheduled",
            "default": false
          },
          "scheduledEffectiveAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledeffectiveat"
          },
          "scheduledTariffId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledtariffid"
          }
        },
        "type": "object",
        "title": "SubscribeOut"
      },
      "SubscribeRequest": {
        "properties": {
          "tariffId": {
            "type": "string",
            "format": "uuid",
            "title": "Tariffid"
          },
          "returnUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Returnurl"
          }
        },
        "type": "object",
        "required": [
          "tariffId"
        ],
        "title": "SubscribeRequest"
      },
      "TariffCreate": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "priceRub": {
            "type": "integer",
            "title": "Pricerub",
            "default": 0
          },
          "periodDays": {
            "type": "integer",
            "title": "Perioddays",
            "default": 30
          },
          "rank": {
            "type": "integer",
            "title": "Rank",
            "default": 0
          },
          "isDefault": {
            "type": "boolean",
            "title": "Isdefault",
            "default": false
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          },
          "sortOrder": {
            "type": "integer",
            "title": "Sortorder",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "code",
          "title"
        ],
        "title": "TariffCreate"
      },
      "TariffOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "priceRub": {
            "type": "integer",
            "title": "Pricerub"
          },
          "periodDays": {
            "type": "integer",
            "title": "Perioddays"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "isDefault": {
            "type": "boolean",
            "title": "Isdefault"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "sortOrder": {
            "type": "integer",
            "title": "Sortorder"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "priceRub",
          "periodDays",
          "rank",
          "isDefault",
          "isActive",
          "sortOrder"
        ],
        "title": "TariffOut"
      },
      "TariffUpdate": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "priceRub": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricerub"
          },
          "periodDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perioddays"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "isDefault": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isdefault"
          },
          "isActive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isactive"
          },
          "sortOrder": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sortorder"
          }
        },
        "type": "object",
        "title": "TariffUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}